pyproject.toml
[tool.poetry]
name = "tree-news"
version = "1.0.0"
description = ""
authors = ["nopareidolia <nopareidolia@protonmail.com>"]
readme = "README.md"
packages = [{include = "tree_news"}]

[tool.poetry.dependencies]
python = "^3.11"
websockets = "^10.4"
discord-py = "^2.2.2"


[tool.poetry.group.dev.dependencies]
pytest = "^7.2.2"
pytest-asyncio = "^0.21.0"
pytest-cov = "^4.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
